@@ -1,19 +1,21 @@
 {
-  "name": "react-autocomplete",
-  "version": "1.8.1",
+  "name": "@connectalum/react-autocomplete",
+  "version": "1.9.0",
   "description": "Accessible, extensible, Autocomplete for React.js",
   "main": "./build/lib/Autocomplete.js",
+  "private": false,
   "files": [
     "build/lib/index.js",
     "build/lib/Autocomplete.js",
+    "build/lib/index.d.ts",
     "dist"
   ],
   "repository": {
     "type": "git",
-    "url": "https://github.com/reactjs/react-autocomplete.git"
+    "url": "https://github.com/DevHeads-Connect/react-autocomplete"
   },
-  "homepage": "https://github.com/reactjs/react-autocomplete",
-  "bugs": "https://github.com/reactjs/react-autocomplete/issues",
+  "homepage": "https://github.com/DevHeads-Connect/react-autocomplete",
+  "bugs": "https://github.com/DevHeads-Connect/react-autocomplete/issues",
   "directories": {
     "example": "examples"
   },
@@ -27,7 +29,7 @@
     "test": "scripty",
     "test:lint": "scripty",
     "test:jest": "scripty",
-    "coverage": "scripty",
+    "coverage": "jest --coverage",
     "start": "SCRIPTY_PARALLEL=true scripty"
   },
   "authors": [
@@ -35,19 +37,19 @@
   ],
   "license": "MIT",
   "devDependencies": {
-    "babel-cli": "^6.5.1",
+    "babel-cli": "^6.26.0",
     "babel-eslint": "6.1.2",
     "babel-jest": "^20.0.3",
-    "babel-preset-es2015": "^6.5.0",
-    "babel-preset-react": "^6.5.0",
+    "babel-preset-es2015": "^6.24.1",
+    "babel-preset-react": "^6.24.1",
     "babel-preset-react-hmre": "^1.1.1",
-    "babel-preset-stage-0": "^6.5.0",
+    "babel-preset-stage-0": "^6.24.1",
     "babelify": "^7.3.0",
     "browserify": "^13.3.0",
-    "browserify-hmr": "^0.3.5",
-    "enzyme": "^2.8.2",
-    "enzyme-to-json": "^1.5.1",
-    "eslint": "^2.13.1",
+    "browserify-hmr": "^0.4.1",
+    "enzyme": "^2.9.1",
+    "enzyme-to-json": "^1.6.0",
+    "eslint": "^8.57.0",
     "eslint-config-rackt": "^1.1.1",
     "eslint-plugin-react": "^6.10.3",
     "exorcist": "^0.4.0",
@@ -55,15 +57,15 @@
     "lodash.flow": "^3.5.0",
     "lodash.map": "^4.6.0",
     "lodash.sortby": "^4.7.0",
-    "react": "^15.5.4",
-    "react-docgen": "^2.15.0",
-    "react-dom": "^15.5.4",
-    "react-test-renderer": "^15.5.4",
-    "scripty": "^1.7.1",
-    "st": "^1.2.0",
-    "uglify-js": "^2.7.5",
-    "watchify": "^3.8.0",
-    "youemdee": "^1.0.0"
+    "react": "^15.7.0",
+    "react-docgen": "^2.21.0",
+    "react-dom": "^15.7.0",
+    "react-test-renderer": "^15.7.0",
+    "scripty": "^1.9.1",
+    "st": "^1.2.2",
+    "uglify-js": "^2.8.29",
+    "watchify": "^3.11.1",
+    "youemdee": "^1.0.1"
   },
   "tags": [
     "react",
@@ -74,7 +76,7 @@
   "keywords": [],
   "dependencies": {
     "dom-scroll-into-view": "1.0.1",
-    "prop-types": "^15.5.10"
+    "prop-types": "^15.8.1"
   },
   "peerDependencies": {
     "react": "^0.14.7 || ^15.0.0-0 || ^16.0.0-0",
