@@ -1,7 +1,7 @@
-const React = require('react')
-const PropTypes = require('prop-types')
-const { findDOMNode } = require('react-dom')
-const scrollIntoView = require('dom-scroll-into-view')
+import React from 'react'
+import PropTypes from 'prop-types'
+import { findDOMNode } from 'react-dom'
+import scrollIntoView from 'dom-scroll-into-view'
 
 const IMPERATIVE_API = [
   'blur',
@@ -16,11 +16,11 @@ const IMPERATIVE_API = [
 
 function getScrollOffset() {
   return {
-    x: (window.pageXOffset !== undefined)
-      ? window.pageXOffset
+    x: (window.scrollX !== undefined)
+      ? window.scrollX
       : (document.documentElement || document.body.parentNode || document.body).scrollLeft,
-    y: (window.pageYOffset !== undefined)
-      ? window.pageYOffset
+    y: (window.scrollY !== undefined)
+      ? window.scrollY
       : (document.documentElement || document.body.parentNode || document.body).scrollTop,
   }
 }
@@ -221,7 +221,7 @@ class Autocomplete extends React.Component {
     this._scrollTimer = null
   }
 
-  componentWillReceiveProps(nextProps) {
+  getDerivedStateFromProps(nextProps) {
     if (this.state.highlightedIndex !== null) {
       this.setState(this.ensureHighlightedIndex)
     }
@@ -597,5 +597,4 @@ class Autocomplete extends React.Component {
   }
 }
 
-module.exports = Autocomplete
-
+export default Autocomplete
