@@ -1,1 +1,3 @@
-module.exports = require('./Autocomplete')
+import Autocomplete from "./Autocomplete";
+
+export default Autocomplete;
